+++ app/controllers/stories_controller.rb
@@ -136,7 +136,8 @@ class StoriesController < ApplicationController
                                 cached_tagged_count
                               end
     @number_of_articles = user_signed_in? ? 5 : SIGNED_OUT_RECORD_COUNT
-    @stories = Articles::Feed.new(number_of_articles: @number_of_articles, tag: @tag, page: @page)
+    @stories = Articles::Feeds::LargeForemExperimental
+      .new(number_of_articles: @number_of_articles, tag: @tag, page: @page)
       .published_articles_by_tag
 
     @stories = @stories.where(approved: true) if @tag_model&.requires_approval
@@ -181,7 +182,7 @@ class StoriesController < ApplicationController
   end
 
   def featured_story
-    @featured_story ||= Articles::Feed.find_featured_story(@stories)
+    @featured_story ||= Articles::Feeds::LargeForemExperimental.find_featured_story(@stories)
   end
 
   def handle_podcast_index
@@ -255,7 +256,7 @@ class StoriesController < ApplicationController
   end
 
   def assign_feed_stories
-    feed = Articles::Feed.new(page: @page, tag: params[:tag])
+    feed = Articles::Feeds::LargeForemExperimental.new(page: @page, tag: params[:tag])
     if params[:timeframe].in?(Timeframer::FILTER_TIMEFRAMES)
       @stories = feed.top_articles_by_timeframe(timeframe: params[:timeframe])
     elsif params[:timeframe] == Timeframer::LATEST_TIMEFRAME
