+++ app/models/site_config.rb
@@ -159,6 +159,7 @@ class SiteConfig < RailsSettings::Base
   # The default font for all users that have not chosen a custom font yet
   field :default_font, type: :string, default: "sans_serif"
   field :primary_brand_color_hex, type: :string, default: "#3b49df"
+  field :feed_strategy, type: :string, default: "basic"
 
   # Broadcast
   field :welcome_notifications_live_at, type: :date
